feat:レビューを受けてリファクタリングしました
