--- src/main/java/domain/GetCountryGreeting.java
@@ -1,8 +1,0 @@
-package domain;
-
-public class GetCountryGreeting {
-    public static  String getCountryGreeting(){
-
-        return "Hoge";
-    }
-}
