+++ src/main/resources/templates/index.html
@@ -10,7 +10,7 @@
 
     <form action="/" method="post">
         <p>国を選んでください。各国の挨拶を確認できます</p>
-        <select name="country_name">
+        <select name="countryName">
             <option value="japan">日本</option>
             <option value="us">アメリカ</option>
             <option value="china">中国</option>
